requirements.LOCAL.txt
aiohttp
aiosignal
async-timeout
attrs
base58
bitarray
certifi
charset-normalizer
currency-symbols
cytoolz
frozenlist
hexbytes
idna
ipfshttpclient
jsonschema
lru-dict
multiaddr
multidict
netaddr
oauthlib
parsimonious
protobuf
pycryptodome
pyrsistent
pyshorteners
python-dateutil
python-dotenv
requests
requests-oauthlib
rlp
six
toolz
tweepy
urllib3
varint
web3
websockets
yarl
eth-utils
eth-abi
eth-account
eth-hash
eth-keyfile
eth-keys
eth-rlp
eth-typing